Add Milestone 2.2: migrate money values from float to Decimal
All monetary fields in PricingConfig, QuoteBreakdown, and QuoteService
now use decimal.Decimal. Non-monetary measurements (print time seconds,
material grams) remain plain numeric types and are converted to Decimal
at the calculation boundary in QuoteService using str() to avoid binary
float representation issues. Decimal values are always constructed from
strings, never floats. Each monetary output field is quantized to
Decimal("0.01") with ROUND_HALF_UP. Built-in profiles updated to use
Decimal string construction. Adds a test proving Decimal avoids the
classic 0.1 + 0.2 float precision error. Tech-debt note in
architecture.md updated to reflect the migration is done.

+++ src/geekcurio_print_manager/models/quote.py
@@ -1,14 +1,15 @@
 from dataclasses import dataclass
+from decimal import Decimal
 
 
 @dataclass(frozen=True, slots=True)
 class QuoteBreakdown:
-    print_time_hours: float
-    print_time_cost: float
-    material_weight_g: float
-    material_cost: float
-    overhead_multiplier: float
-    markup_percentage: float
-    subtotal: float
-    markup_amount: float
-    total: float
+    print_time_hours: float       # measurement — stays float
+    print_time_cost: Decimal
+    material_weight_g: float      # measurement — stays float
+    material_cost: Decimal
+    overhead_multiplier: Decimal  # carried from PricingConfig
+    markup_percentage: Decimal    # carried from PricingConfig
+    subtotal: Decimal
+    markup_amount: Decimal
+    total: Decimal
